+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Acronym identification fixes.
 
 ## [0.2.2](https://github.com/zobweyt/textcase/compare/0.2.1...0.2.2) (2025-04-12)
 
-Performance and documentation enhancements
+Performance and documentation enhancements.
 
 ### <!-- 2 -->Performance <!-- omit in toc -->